This is Mark Bird's EMDeveloperTechnicalTest solution.
Sorry about the Identity package - I brought down too much!

Main task and 3/4 sub tasks completed.

The  register and login works (see screenshot in email) but I couldn't get it to hide the payment page on unauthorised access. I think I may have used an older Identity login / register framework but got it logging in and out and the user in the database.

Added the API key to headers in the front end startup and this is analysed by the backend in the APIKeyAttribute file.

Finally client side and server side validation implemented, however I kept getting a 400 error because the datetime input was sending a string back to the server and it expects a datetime object (or null), I didn't have time to sort that.